[ottf,console] Move code from ottf_console.c to device-specific files
No code was changed, a few extern declarations were added to the internal
header.
